feat: l'ambiente di riferimento è ora test comuni
free test dismesso
https://github.com/italia/anpr/issues/3403

Aggiunto CHANGELOG.md
