+++ CHANGELOG.md
@@ -1,0 +1,27 @@
+# Changelog
+
+Tutte le modifiche significative a questo progetto verranno documentate in questo file.
+
+Il formato è basato su [Tenere un changelog](https://keepachangelog.com/en/1.0.0/),
+e il progetto si conforma al [Versionamento semantico](https://semver.org/spec/v2.0.0.html).
+
+## [Non rilasciato]
+
+### Aggiunto
+
+- CHANGELOG.md (questo file)
+- trust con certificato wstest.anpr.interno.it aggiornato
+
+### Modificato
+
+- l'ambiente di riferimento è ora test comuni
+
+### Rimosso
+
+- supporto ambiente free test [italia/anpr #3403](https://github.com/italia/anpr/issues/3403).
+
+## [1.3.0] - XXXX-XX-XX
+
+### Aggiunto
+
+- fork da <https://github.com/italia/anpr-client-example>
