+++ request/6001/6001_888013_TEST.req
@@ -13,7 +13,7 @@
 
     <tipoOperazione>C</tipoOperazione>
 
-    <tipoInvio>FREE</tipoInvio>
+    <tipoInvio>TEST</tipoInvio>
 
     <dataDefinizionePratica>2018-05-29</dataDefinizionePratica>
