+++ src/main/java/it/interno/anpr/config/EnvironmentHandler.java
@@ -7,12 +7,11 @@ import java.util.List;
 public class EnvironmentHandler {
 
 	public static final String VAL = "VAL";
-	public static final String TEST = "TEST";
 	public static final String PRE = "PRE";
-	public static final String FREE = "FREE";
+	public static final String TEST = "TEST";
 	public static final String PROD = "PROD";
 
-	private static List<String> envList = new ArrayList<String>( Arrays.asList(VAL, TEST, PRE, FREE, PROD));
+	private static List<String> envList = new ArrayList<String>( Arrays.asList(VAL, TEST, PRE, TEST, PROD));
 	private static String env;
 
 	public static String getEnv() {
