+++ src/main/resources/config/Example_TEST_Keystore.properties
@@ -1,16 +1,16 @@
 # Properties WSS4J - Keystore del Comune
-org.apache.ws.security.crypto.merlin.keystore.file=keystore/FREE/<Nome file keystore>.p12
+org.apache.ws.security.crypto.merlin.keystore.file=keystore/TEST/<Nome file keystore>.p12
 org.apache.ws.security.crypto.merlin.keystore.password=<Password>
 org.apache.ws.security.crypto.merlin.keystore.type=pkcs12
 org.apache.ws.security.crypto.merlin.keystore.alias=<ID Postazione>
 
 # Properties internal
-ID_OPERATORE=<Username web (USERSFREETEST)>
+ID_OPERATORE=<Username web (Utente test comuni)>
 ID_SEDE=<Codice Comune>
 ID_POSTAZIONE=<ID Postazione>
 ID_APPLICAZIONE=12345
 ID_TRANSAZIONE=99999
-PATH_KEYSTORE_POSTAZIONE=keystore/FREE/<Nome file keystore>.p12
+PATH_KEYSTORE_POSTAZIONE=keystore/TEST/<Nome file keystore>.p12
 PASS_KEYSTORE_POSTAZIONE=<Password>
 ALIAS_KEYSTORE_POSTAZIONE=<ID Postazione>
-URL_WS=https://wsfree.anpr.interno.it
+URL_WS=https://wstest.anpr.interno.it
